+++ pandapower/pf/runpp_3ph.py
@@ -195,7 +195,7 @@ def runpp_3ph(net, calculate_voltage_angles=True, init="auto",
             - transformer equivalent models
             - "t" - transformer is modeled as equivalent with the T-model.
             - "pi" - This is not recommended, since it is less exact than the T-model.
-                So, for three phase load flow, its not implemented
+                So, for three phase load flow, it's not implemented
 
 
         **trafo_loading** (str, "current") - mode of calculation for
@@ -224,12 +224,11 @@ def runpp_3ph(net, calculate_voltage_angles=True, init="auto",
 
             Note: enforce_q_lims only works if algorithm="nr"!
 
-
         **check_connectivity** (bool, True) - Perform an extra connectivity
         test after the conversion from pandapower to PYPOWER.
 
             If True, an extra connectivity test based on SciPy Compressed
-            Sparse Graph Routines is perfomed. If check finds unsupplied buses,
+            Sparse Graph Routines is performed. If check finds unsupplied buses,
             they are set out of service in the ppc
 
         **voltage_depend_loads** (bool, True)
@@ -383,7 +382,7 @@ def runpp_3ph(net, calculate_voltage_angles=True, init="auto",
                      trafo_model=trafo_model, check_connectivity=check_connectivity,
                      mode=mode, switch_rx_ratio=switch_rx_ratio,
                      init_vm_pu=init, init_va_degree=init,
-                     enforce_q_lims=enforce_q_lims, recycle=None,
+                     enforce_q_lims=enforce_q_lims, enforce_p_lims=False, recycle=None,
                      voltage_depend_loads=False, delta=delta_q,
                      neglect_open_switch_branches=neglect_open_switch_branches
                      )