+++ pandapower/timeseries/output_writer.py
@@ -402,7 +402,7 @@ class OutputWriter(JSONSerializableClass):
             # if no ppc is in net-> create one
             options = dict(algorithm='nr', calculate_voltage_angles=True, init="auto",
                            max_iteration="auto", tolerance_mva=1e-8, trafo_model="t",
-                           trafo_loading="current", enforce_q_lims=False, check_connectivity=True,
+                           trafo_loading="current", enforce_p_lims=False, enforce_q_lims=False, check_connectivity=True,
                            voltage_depend_loads=True, consider_line_temperature=False)
             _init_runpp_options(net, **options)
             ppc, _ = _pd2ppc(net)